[FE-1266] AggregateMetrics component
